keep doubles for all the solar elevation/azimuth/distance stuff everywhere chop it to 1 decimal of precision ofr elevation/azimuth in the snapForTelemtry t_* stuff
standardize on print 7 digits of precision for lat/lon?

+++ tracker/solar2_functions.cpp
@@ -144,7 +144,7 @@ void solarZenithAndAzimuthAngle2(double *sza, double *saa, double longitude, dou
 //***********************************************************************
 
 // we don't have distance for this fast algo
-void calcSolarElevation2(int *solarElevation, int *solarAzimuth, int *solarDistance) {
+void calcSolarElevation2(double *solarElevation, double *solarAzimuth, double *solarDistance) {
     V1_print(F("calcSolarElevation2 START" EOL));
 
     // we check before setting rtc. assuming these are all valid ranges
@@ -239,12 +239,12 @@ void calcSolarElevation2(int *solarElevation, int *solarAzimuth, int *solarDista
         elevation, altitude);
 
     // returns degrees, not radians
-    // solarElevation: can this be negative?. decimal, integer accuracy
-    // solarAzimuth:   can this be negative?. decimal, integer accuracy
-    // solarDistance:  can this be negative? maybe error case. kilometers.
-    double solarElevation_here = (int)(sza);
-    double solarAzimuth_here   = (int)(saa);
-    double solarDistance_here  = (int)(0);
+    // solarElevation: can this be negative?. double
+    // solarAzimuth:   can this be negative?. double
+    // solarDistance:  can this be negative? double
+    double solarElevation_here = sza;
+    double solarAzimuth_here   = saa;
+    double solarDistance_here  = 0;
 
     if (badSolar) {
         solarElevation_here = 0.0;
@@ -252,9 +252,9 @@ void calcSolarElevation2(int *solarElevation, int *solarAzimuth, int *solarDista
         solarDistance_here  = 0.0;
     }
 
-    *solarElevation = (int)solarElevation_here;
-    *solarAzimuth   = (int)solarAzimuth_here;
-    *solarDistance  = (int)solarDistance_here;
+    *solarElevation = solarElevation_here;
+    *solarAzimuth   = solarAzimuth_here;
+    *solarDistance  = solarDistance_here;
 
     V1_print(F("calcSolarElevation2 END" EOL));
 }
