+++ tracker/solar4_functions.h
@@ -10,7 +10,7 @@
 // #define INVALID_VALUE 999999999.999999d
 
 void solarZenithAndAzimuthAngle4(double *sza, double *saa, double longitude, double latitude, time_t timeStamp);
-void calcSolarElevation4(int *solarElevation, int *solarAzimuth, int *solarDistance);
+void calcSolarElevation4(double *solarElevation, double *solarAzimuth, double *solarDistance);
 time_t getEpochTime4();
 
 // from solarfunctions.cpp originally
